Create integration test base class
